requirements.txt
google-generativeai==0.3.2
groq==0.3.0
PyPDF2==3.0.1
python-dotenv==1.0.0
numpy==1.24.3
pandas==2.1.4
nltk==3.8.1
sentence-transformers==2.2.2
scikit-learn==1.3.0
tqdm==4.66.1
transformers==4.36.0
torch==2.1.0
streamlit==1.32.0
plotly==5.18.0
alpha_vantage==2.3.1
yfinance==0.2.36
streamlit-option-menu==0.3.12
streamlit-extras==0.4.0
streamlit-lottie==0.0.5 